requirements.txt
PyYAML
Appium-Python-Client
